Refactor, split files.
