+++ arwendebug/messagesQuery.go
@@ -1,0 +1,11 @@
+package arwendebug
+
+// QueryRequest is a CLI / REST request message
+type QueryRequest struct {
+	RunRequest
+}
+
+// QueryResponse is a CLI / REST response message
+type QueryResponse struct {
+	ContractResponseBase
+}
