+++ arwendebug/messagesRun.go
@@ -33,13 +33,3 @@ func (request *RunRequest) digest() error {
 type RunResponse struct {
 	ContractResponseBase
 }
-
-// QueryRequest is a CLI / REST request message
-type QueryRequest struct {
-	RunRequest
-}
-
-// QueryResponse is a CLI / REST response message
-type QueryResponse struct {
-	ContractResponseBase
-}
